explorer/requirements.txt
functions-framework
google-cloud-storage
opencv-python
requests